% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClimateFile.R
\name{createClimateFile}
\alias{createClimateFile}
\title{createClimateFile}
\usage{
createClimateFile(latitude, longitude, start_date, end_date, placename)
}
\arguments{
\item{latitude}{Latitude for modeling location}

\item{longitude}{Longitude for modeling location}

\item{start_date}{Starting date for model run}

\item{end_date}{Ending date for model run}

\item{placename}{Text string for model location, used in file naming}
}
\description{
This function takes in the relevant parameters needed for making API calls and proceeds to
create the 4 files necessary for use in the AquaCrop software. The 4 files will be saved to the current
R working directory, after which they may be copied into the DATA folder within your AquaCrop installation location.
From there you should now be able to run the AquaCrop software and select the climate file by finding the placename, specified in this
function, in the list of climate files within the program.
}
